requirements.txt
faker
kafka-python
tinybird-cli